refactor: Moves shared function into separate file
- Moves functions to generate test data into conftest.py
- Moves function to load data from json file to utils.py
