+++ analytics/tests/datasets/test_sprint_board.py
@@ -1,63 +1,12 @@
-import json
-
 import pandas as pd
 import pytest
 
 from analytics.datasets.sprint_board import SprintBoard
-
-
-def json_issue_row(
-    issue: int,
-    created_at: str = "2023-11-01T00:00:00Z",
-    closed_at: str = "2023-11-01T00:00:00Z",
-) -> dict:
-    """Generate a row of JSON issue data for testing"""
-    return {
-        "closedAt": closed_at,
-        "createdAt": created_at,
-        "number": issue,
-    }
-
-
-def json_sprint_row(
-    issue: int,
-    parent_number: int = -99,
-    sprint_name: str = "Sprint 1",
-    sprint_date: str = "2023-11-01",
-    status: str = "Done",
-    story_points: int = 5,
-) -> dict:
-    """Generate a row of JSON sprint data for testing"""
-    return {
-        "assignees": ["mickeymouse"],
-        "content": {
-            "type": "Issue",
-            "body": f"Description of test issue {issue}",
-            "title": f"Test issue {issue}",
-            "number": issue,
-            "repository": "HHS/simpler-grants-gov",
-            "url": f"https://github.com/HHS/simpler-grants-gov/issues/{issue}",
-        },
-        "id": "PVTI_lADOABZxns4ASDf3zgJhmCk",
-        "labels": ["topic: infra", "project: grants.gov"],
-        "linked pull requests": [],
-        "milestone": {
-            "title": "Sample milestone",
-            "description": f"30k ft deliverable: #{parent_number}",
-            "dueOn": "2023-10-20T00:00:00Z",
-        },
-        "repository": "https://github.com/HHS/simpler-grants-gov",
-        "sprint": {"title": sprint_name, "startDate": sprint_date, "duration": 14},
-        "status": status,
-        "story Points": story_points,
-        "title": "Test issue 1",
-    }
-
-
-def write_test_data_to_file(data: dict, output_file: str):
-    """"""
-    with open(output_file, "w") as f:
-        f.write(json.dumps(data))
+from tests.conftest import (
+    json_issue_row,
+    json_sprint_row,
+    write_test_data_to_file,
+)
 
 
 class TestSprintBoard:
